Case 135 - FR - Satanée Rétro Compatibilité
I couldn't find a translation for "Ass-backwards"
(is it about riding an ass backwards? is it about half-assed job?)
so I put a pun of my own on "vade retro satanas".
Cheers!

PS: I've had case 128 in my drafts for a long time... but it's so full of graphic rendering references that I'll have a friend review it before, if he's willing to.