applicant-1 - FR - L'Ermite
OK, a few comments here.
First, this is significantly longer but I will of course translate The Applicant.

Second, I took the habit of googling titles (and cryptic idioms) hoping I'll catch references.
It so happens that La Candidate (The Applicant) is a porn film, so I renamed the series La Candidature (The Application).

Then, we have "sacked" in FR albeit with a slightly different meaning. It still works.
But it was hard translating "fired", and impossible "canned" - I didn't even find a dictionary to help on this one.
So I hesitated to use "placardiser" (put someone in a closet / a dead-end job) but chose "changer de boîte" (change box / change company).

That's all for now folks.
Cheers!